requirements.txt
pymongo
python-dotenv
inquirerpy
